[FIX] Wizard with Multiple emails: Generate the correct attachment to each record, with the correct lang. Link to the partner event is it is defined.
Remeber: Last commit (Merge of Sharoon) adds this improvement:
Improved poweremail send wizard when multiple emails are sent giving more flexibility to the user:
When multiple emails are sent, the wizard shows all the email fields filled with the template content and the user can change them.
Next step, the previous values are used as the template to create and send the multiple emails.